requirements.txt
torch>=2.0.0
smolagents
transformers>=4.30.0
datasets>=2.12.0
openai
python-dotenv>=1.0.0